requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-modeltranslation==0.19.19
django-unfold==0.73.1
djangorestframework==3.16.1
environs==14.5.0
idna==3.11
marshmallow==4.1.1
mysqlclient==2.2.7
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.6.2
